requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
discord.py==1.7.3
et-xmlfile==1.0.1
idna==3.1
jdcal==1.4.1
multidict==5.1.0
openpyxl==3.0.5
python-dotenv==0.19.0
typing-extensions==3.10.0.0
yarl==1.6.3
